Move detect_sources and make_source_mask to segmentation package
